+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Pre release
 ### Implemented enhancements
+- GB2-918 - Added Internet Usage and CPI data for PopulationByCountry endpoint
 - GBAU-61 - Added company endpoint for consumption by activity stream
 - GP2-510 - Added Suggested Countries model and relevant data
 - GP-96 - HS Codes saved on Country